Testing lesson 5

+++ app/src/main/java/come/geekbrains/vitekm/searchmovies/model/repository/RepositoryImpl.kt
@@ -20,7 +20,7 @@ class RepositoryImpl : Repository {
                     i.fullTitle.toString(),
                     i.imDbRating.toString(),
                     i.year,
-                    i.image.toInt(),
+                    i.image,
                     i.crew.toString()
                 )
                 result.add(filmItem)
